fix: fields 'size' and 'hasAttachments' not returned to client
refs conjoon/php-ms-imapuser#43
