+++ lib/tests/Http/V0/Query/MessageItem/IndexRequestQueryTranslatorTest.php
@@ -418,6 +418,8 @@ class IndexRequestQueryTranslatorTest extends TestCase
             "charset" => true,
             "references" => true,
             "messageId" => true,
+            "size" => true,
+            "hasAttachments" => true,
             "html" =>  ["length" => 200, "trimApi" => true, "precedence" => true],
             "plain" => ["length" => 200, "trimApi" => true]
         ];
@@ -444,7 +446,9 @@ class IndexRequestQueryTranslatorTest extends TestCase
             "charset",
             "references",
             "messageId",
-            "previewText"
+            "previewText",
+            "size",
+            "hasAttachments"
         ];
     }
 }
